Movement between rooms
Implementation should add ability to move between rooms. At the very least, it tells us that the doors are all locked when an attempt is made (Which should be changed: No monsters in first room).

Additionally, I commented out the lines which allowed the player to click on a door to transfer rooms.